Big refactors to table_io tests
