+++ daft/table/table_io.py
@@ -125,8 +125,6 @@ def read_csv_with_schema(
         csv_options (vPartitionParseCSVOptions, optional): options for reading the CSV file. Defaults to vPartitionParseCSVOptions().
         read_options (vPartitionReadOptions, optional): options for reading the Table. Defaults to vPartitionReadOptions().
     """
-    csv_is_headerless = csv_options.header_index is None
-
     with _get_file(file) as f:
 
         if read_options.num_rows is not None:
@@ -148,17 +146,14 @@ def read_csv_with_schema(
                 delimiter=csv_options.delimiter,
             ),
             read_options=pacsv.ReadOptions(
-                # If CSV has no header, we assume the provided schema's (ordered) field names as the column names
-                # Otherwise, we will read the headers from the CSV (indicated with column_names=None and autogenerate_column_names=False)
-                column_names=schema.column_names() if csv_is_headerless else None,
-                autogenerate_column_names=False,
-                skip_rows=(0 if csv_is_headerless else csv_options.header_index),
-            ),
-            convert_options=pacsv.ConvertOptions(
-                include_columns=read_options.column_names,
-                include_missing_columns=True,
+                # Use the provided schema's field names as the column names, and skip parsing headers entirely
+                column_names=schema.column_names(),
+                skip_rows=(0 if csv_options.header_index is None else csv_options.header_index + 1),
             ),
         )
+
+    if read_options.column_names is not None:
+        table = table.select(read_options.column_names)
 
     return Table.from_arrow(table)
 
@@ -169,7 +164,7 @@ def infer_schema_csv(
     csv_options: vPartitionParseCSVOptions = vPartitionParseCSVOptions(),
     read_options: vPartitionReadOptions = vPartitionReadOptions(),
 ) -> Schema:
-    """Reads a Table from a CSV file
+    """Reads a Schema from a CSV file
 
     Args:
         file (str | IO): either a file-like object or a string file path (potentially prefixed with a protocol such as "s3://")
@@ -179,7 +174,7 @@ def infer_schema_csv(
         read_options (vPartitionReadOptions, optional): Options for reading the file
 
     Returns:
-        Table: Parsed Table from CSV
+        Schema: Inferred Schema from the CSV
     """
     # Have PyArrow generate the column names if the CSV has no header and no column names were provided
     pyarrow_autogenerate_column_names = (csv_options.header_index is None) and (override_column_names is None)
@@ -194,9 +189,7 @@ def infer_schema_csv(
         if read_options.num_rows is not None:
             num_rows_to_read = (
                 # Extra rows before the header
-                0
-                if csv_options.header_index is None
-                else csv_options.header_index
+                (0 if csv_options.header_index is None else csv_options.header_index)
                 # Header row
                 + (1 if csv_options.header_index is not None else 0)
                 # Actual number of data rows to read
